requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blessings==1.7
blinker==1.4
bokeh==2.3.3
boto3==1.4.8
botocore==1.8.50
branca==0.4.2
cachetools==4.2.2
Cartopy @ file:///H:/Documents%20and%20Settings/Personal/My%20Received%20Files/Cartopy-0.18.0-cp36-cp36m-win_amd64.whl
certifi==2021.5.30
cffi==1.14.5
cfgrib==0.9.9.0
cftime==1.5.0
chardet==4.0.0
click==7.1.2
cloudpickle==1.4.1
cmocean==2.0
colorama==0.4.4
colorful==0.5.4
contextvars==2.4
convertdate==2.3.2
custom-inherit==2.3.2
cycler==0.10.0
Cython==0.29.24
dask==2.19.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
distributed==2.19.0
docutils==0.17.1
eccodes==1.3.3
ecmwflibs==0.3.7
entrypoints==0.3
eofs==1.4.0
ffmpeg==1.4
filelock==3.0.12
findlibs==0.0.2
Flask==2.0.1
folium==0.12.1
geoplotlib==0.3.2
geos==0.2.3
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.30.0
google-auth==1.32.1
googleapis-common-protos==1.53.0
gpustat==0.6.0
graphviz==0.17
grpcio==1.38.1
HeapDict==1.0.1
hijri-converter==2.2.0
hiredis==2.0.0
holidays==0.11.2
idna==2.10
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==4.6.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.2.0
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
lab==6.3
lxml==4.6.3
markovchain==0.2.5
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
modin==0.8.3
msgpack==1.0.2
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nco==1.0.0
nest-asyncio==1.5.1
netCDF4==1.5.6
networkx==2.5.1
nio==3.4.2
notebook==6.4.0
numexpr==2.7.3
numpy==1.19.5
nvidia-ml-py3==7.352.0
opencensus==0.7.13
opencensus-context==0.1.2
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.0
plotly==5.1.0
plumbum==1.7.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
py-spy==0.3.7
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycmap==0.1.10
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2
pydot==1.4.2
pydotplus==2.0.2
pyeccodes==0.1.1
Pygments==2.9.0
PyMeeus==0.5.11
pyparsing==2.4.7
pypiwin32==223
pyproj==3.0.1
pyrsistent==0.18.0
pyshp==2.1.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
ray==1.1.0
redis==3.5.3
requests==2.25.1
rpyc==4.1.5
rsa==4.7.2
s3transfer==0.1.13
safepickle==0.2.0
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.7.1
Shapely==1.7.1
simplejson==3.17.2
six==1.16.0
sklearn==0.0
sklearn-xarray==0.4.0
smmap==4.0.0
sortedcontainers==2.4.0
streamlit==0.86.0
tables==3.6.1
tblib==1.7.0
tenacity==8.0.1
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.2
traitlets==4.3.3
txt2tags==3.7
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xarray==0.16.2
yarl==1.6.3
zict==2.0.0
zipp==3.5.0
